Extracted test base class
